refactor(rts-core): hoist extract_symbols to pub(crate) fn (B0)
Moves the body of `CodebaseAnalyzer::extract_symbols` + its 17
per-language helpers (`extract_rust_symbols`, doc-comment scanners,
etc.) out of the `impl CodebaseAnalyzer` block and into a new
`pub(crate) mod extraction` as free functions. None of the helpers
actually used `&self` — the conversion is mechanical.

This is B0 of PR-B from the pre-pivot cleanup plan. It unblocks
B1's `parse_content` facade, which calls `extraction::extract_symbols`
directly without instantiating a stateful analyzer. The
`CodebaseAnalyzer::analyze_content` method now delegates to the
free function for the daemon's existing callsite — behavior is
unchanged, but the extraction logic is no longer entangled with
the analyzer struct that PR-B (B3) is about to delete.

- `crates/rts-core/src/extraction.rs` (new, 1283 LoC): 18 free
  functions, `extract_symbols` is the single dispatch entry point
- `crates/rts-core/src/lib.rs`: register `pub(crate) mod extraction`
- `crates/rts-core/src/analyzer.rs`: replace 3 `self.extract_symbols`
  callsites with `crate::extraction::extract_symbols`; delete the
  1497 lines of now-duplicated helpers; drop unused `SyntaxTree` import

`cargo test -p rust_tree_sitter --lib` passes (198/198).
